requirements.txt
Flask==2.1.2
requests==2.27.1
markdown==3.3.6
Werkzeug==2.0.3
python-dotenv==0.19.2